nfapi ipsec: Use default statistics settings on outbound policies
NF IPSec layer is using manually added outbound policy entries for
orphan IPSec policies, but it was not using the same statistics
settings like it did for the created SAs. This was causing the
function nf_ipsec_global_stats_get to fail when it was trying to
get the statistics for such policy entries.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ib7559620c985066fa6331da1691bf63952318253
Reviewed-on: http://git.am.freescale.net:8181/36043
Reviewed-by: Anca Jeanina Dragusin <AncaJeanina.Dragusin@freescale.com>