nfapi ipsec: Initialize IP address masks to zero before updating them
The function "set_ip_addr_mask" is used to update the masks used
for IP addresses according to the requested IP address prefix. This
function fills in bits of 1 where needed until the address prefix is
covered. However, it does not check that the remaining bits are
cleared therefore one needs to initialize the IP address mask to zero
before calling "set_ip_addr_mask" to make sure a consistent mask is
created.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I1952047c6a292823232a3c541c19bacc35b3fd3e
Reviewed-on: http://git.am.freescale.net:8181/36071
Reviewed-by: Pinghua An <pinghua.an@freescale.com>